requirements.txt
numpy
opencv-python
pymupdf
requests
scipy
pillow
playwright
matplotlib
